refactor(tests): split test_cli_doctor.py into a focused cli_doctor package (#443)
The doctor-command tests were a single 998-line module spanning ten
distinct concerns (show_info, claude-plugin lookup, overlay-skill
discovery, symlink repair, editable sanity, repo discovery,
make_editable, IntrospectionHelpers, the CliRunner doctor-check
command, and the pure-function helpers).

Split it verbatim into tests/cli_doctor/, one module per concern, with
the shared TOML/home-staging and fake-entry-point helpers hoisted into
_shared.py so each focused module imports them instead of redefining
them. No conftest.py is needed: every fixture used is a pytest builtin
(tmp_path/monkeypatch/capsys). Test bodies are AST-identical to the
original; only import paths were adjusted and the module-level
CliRunner() relocated to its single remaining consumer.

Same 64 tests collected, same Class::method multiset, all green.
Pure test-only move — zero production changes.

Relates-to #443
